backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1
email-validator>=2.0.0
python-multipart==0.0.6
google-cloud-texttospeech==2.14.0
pillow==10.1.0
